moved compiler updates to helper file, extreme forcing of tests to wait for compilation and not use inline
