+++ test/helper.js
@@ -1,5 +1,8 @@
 'use strict';
 
+/* eslint-disable
+  no-undefined
+*/
 const webpack = require('webpack');
 const Server = require('../lib/Server');
 
@@ -9,9 +12,25 @@ module.exports = {
   // start server, returning the full setup of the server
   // (both the server and the compiler)
   startFullSetup(config, options, done) {
-    // eslint-disable-next-line no-undefined
     if (options.quiet === undefined) {
       options.quiet = true;
+    }
+    // originally, inline was not working by default for tests with the API
+    // if you need to test inline, I think it should be set explicitly,
+    // rather than expecting it to be defaulted to
+    // (the only test that relied on inline before this point was Client.test.js)
+    if (
+      options.inline === undefined &&
+      options.hot === undefined &&
+      options.hotOnly === undefined
+    ) {
+      options.inline = false;
+    }
+    // defaulting to this will hopefully help with problems on OSX in tests
+    if (options.watchOptions === undefined) {
+      options.watchOptions = {
+        poll: true,
+      };
     }
     const compiler = webpack(config);
     server = new Server(compiler, options);
@@ -47,6 +66,13 @@ module.exports = {
     return this.startAwaitingCompilationFullSetup(config, options, done).server;
   },
   start(config, options, done) {
+    // I suspect that almost all tests need to wait for compilation to
+    // finish, because not doing so leaves open handles for jest,
+    // in the case where a compilation didn't finish before destroying
+    // the server and moving on
+    return this.startAwaitingCompilation(config, options, done);
+  },
+  startBeforeCompilation(config, options, done) {
     return this.startFullSetup(config, options, done).server;
   },
   close(done) {
